Strip client-specific properties from annotations before exporting
