+++ src/sidebar/services/annotations-exporter.ts
@@ -1,4 +1,5 @@
-import type { Annotation } from '../../types/api';
+import type { APIAnnotationData } from '../../types/api';
+import { stripInternalProperties } from '../helpers/strip-internal-properties';
 import { VersionData } from '../helpers/version-data';
 import type { SidebarStore } from '../store';
 
@@ -6,7 +7,7 @@ export type ExportContent = {
   export_date: string;
   export_userid: string;
   client_version: string;
-  annotations: Annotation[];
+  annotations: APIAnnotationData[];
 };
 
 /**
@@ -26,7 +27,9 @@ export class AnnotationsExporter {
    */
   buildExportContent(now = new Date()): ExportContent {
     const profile = this._store.profile();
-    const annotations = this._store.allAnnotations();
+    const annotations = this._store
+      .allAnnotations()
+      .map(stripInternalProperties) as APIAnnotationData[];
     const versionData = new VersionData(profile, []);
 
     return {
